Promote small integer types to single rather than double precision (#278)
This is a performance-related PR that will result in casting of small integer dtypes (8 and 16-bit) to 32-bit floats rather than the current use of 64-bit floats. The current behavior is consistent with scikit-image, although I have raised an issue there to potentially change to the behavior proposed here: https://github.com/scikit-image/scikit-image/issues/6310. 

The changes required are quite small with the key one being the change to the `new_float_type` dict that gets used when promoting dtypes to a floating point type. Most other changes are in the tests where we sometimes have to bump up the tolerance when computations that were previously in double precision now get run in single precision instead.

I marked this as **breaking** because the output of various floating-point functions for integer inputs may now be `cp.float32` in cases where it was previously `cp.float64`.

Authors:
  - Gregory Lee (https://github.com/grlee77)

Approvers:
  - Gigon Bae (https://github.com/gigony)
  - https://github.com/jakirkham

URL: https://github.com/rapidsai/cucim/pull/278

+++ python/cucim/src/cucim/skimage/feature/template.py
@@ -134,7 +134,9 @@ def match_template(image, template, pad_input=False, mode='constant',
     image_shape = image.shape
 
     float_dtype = _supported_float_type(image.dtype)
-    image = image.astype(float_dtype, copy=False)
+
+    # Note: keep image in float64 for accuracy of cumsum operations, etc.
+    image = image.astype(cp.float64, copy=False)
     template = template.astype(float_dtype, copy=False)
 
     pad_width = tuple((width, width) for width in template.shape)
@@ -153,11 +155,12 @@ def match_template(image, template, pad_input=False, mode='constant',
         image_window_sum = _window_sum_3d(image, template.shape)
         image_window_sum2 = _window_sum_3d(image * image, template.shape)
 
-    template_mean = template.mean()
+    # perform mean and sum in float64 for accuracy
+    template_mean = template.mean(dtype=cp.float64)
     template_volume = _misc.prod(template.shape)
     template_ssd = template - template_mean
     template_ssd *= template_ssd
-    template_ssd = cp.sum(template_ssd)
+    template_ssd = cp.sum(template_ssd, dtype=cp.float64)
 
     if image.ndim == 2:
         xcorr = signal.fftconvolve(image, template[::-1, ::-1],
